Move authorship-related code to its own files
